chore: extract convoy issue tracking into convoy_issues.go
Split convoy.go by moving issue-level functions (extractIssueID,
getTrackedIssueStatus, refreshTrackedStatus) and the trackedStatus
type into a dedicated convoy_issues.go file for better organization.
